+++ batch_scripts/linux/installation/install_pip_python_packages.sh
@@ -28,7 +28,7 @@ if [ $TCAPY_PYTHON_ENV_TYPE == "virtualenv" ]; then
         mysql-connector-python==8.0.19 \
         IPython==7.14.0 chartpy==0.1.8 findatapy==0.1.12 dash-auth==1.3.2 cufflinks==0.17.3 plotly==4.9.0 chart-studio==1.1.0 \
         dash==1.12.0 dash-html-components==1.0.3 dash-core-components==1.10.0 dash-table==4.7.0 jupyter-dash==0.2.1 dtale==1.8.1 kaleido \
-        dtale==1.8.1 \
+        dtale==1.8.1 dash-bootstrap-components==0.10.3 \
         qpython==2.0.0 influxdb==5.2.3 \
         Flask-Session==0.3.1 \
         celery==4.4.0 pytest-tap kombu==4.6.7 python-memcached==1.59 numba==0.48.0 vispy==0.6.4 jinja2==2.11.2 \
@@ -59,7 +59,7 @@ elif [ $TCAPY_PYTHON_ENV_TYPE == "conda" ] && [ $CONDA_FROM_YAML == 0 ]; then
     # for flash recording of session variables
     # to allow celery to use Redis
     pip install mysql-connector-python==8.0.19 chartpy==0.1.8 findatapy==0.1.12 dash-auth==1.3.2 cufflinks==0.17.3 plotly==4.9.0 \
-        chart-studio==1.1.0 kaleido \
+        chart-studio==1.1.0 kaleido dash-bootstrap-components==0.10.3 \
         dash==1.12.0 dash-html-components==1.0.3 dash-core-components==1.10.0 dash-table==4.7.0 jupyter-dash==0.2.1 dtale==1.8.1 \
         qpython==2.0.0 influxdb==5.2.3 \
         Flask-Session==0.3.1 \